# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is Todennus - An Open ID Connect and OAuth2 Provider",
        "title": "Todennus API Endpoints",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/oauth2_clients": {
            "post": {
                "description": "Create an new OAuth2 Client. If the `is_confidential` field is true, a secret is issued. Please carefully store this secret in a confidential place. This secret will never be retrieved by anyway. \u003cbr\u003e\nRequire scope `create:client`.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OAuth2 Client"
                ],
                "summary": "Create oauth2 client",
                "parameters": [
                    {
                        "description": "Client Information",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.OAuth2ClientCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Create client successfully",
                        "schema": {
                            "$ref": "#/definitions/standard.SwaggerSuccessResponse-dto_OAuth2ClientCreateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/standard.SwaggerBadRequestErrorResponse"
                        }
                    }
                }
            }
        },
        "/oauth2_clients/first": {
            "post": {
                "description": "Create the first OAuth2 Client (always a confidential Client). \u003cbr\u003e\nWhy this API? When todennus is started, there is no existed Client, we don't have any flow to authenticate a user (all authentication flows require a Client). This API is only valid if there is no existing Client and the user is administrator.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OAuth2 Client"
                ],
                "summary": "Create the first oauth2 client",
                "parameters": [
                    {
                        "description": "Client Information",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.OAuth2ClientCreateFirstRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Create client successfully",
                        "schema": {
                            "$ref": "#/definitions/standard.SwaggerSuccessResponse-dto_OAuth2ClientCreateFirstResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/standard.SwaggerBadRequestErrorResponse"
                        }
                    },
                    "401": {
                        "description": "unauthorized",
                        "schema": {
                            "$ref": "#/definitions/standard.SwaggerUnauthorizedErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/standard.SwaggerForbiddenErrorResponse"
                        }
                    },
                    "404": {
                        "description": "API not found",
                        "schema": {
                            "$ref": "#/definitions/standard.SwaggerNotFoundErrorResponse"
                        }
                    }
                }
            }
        },
        "/oauth2_clients/{client_id}": {
            "get": {
                "description": "Get OAuth2 Client information by ClientID. \u003cbr\u003e\nRequire scope `read:client.allowed_scope` to get the allowed scope. \u003cbr\u003e\nRequire scope `read:client.owner` to get Owner ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OAuth2 Client"
                ],
                "summary": "Get oauth2 client by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ClientID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get client successfully",
                        "schema": {
                            "$ref": "#/definitions/standard.SwaggerSuccessResponse-dto_OAuth2ClientGetResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/standard.SwaggerBadRequestErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "$ref": "#/definitions/standard.SwaggerNotFoundErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.OAuth2ClientCreateFirstRequest": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string",
                    "example": "First Client"
                },
                "password": {
                    "type": "string",
                    "example": "s3Cr3tP@ssW0rD"
                },
                "username": {
                    "type": "string",
                    "example": "huykingsofm"
                }
            }
        },
        "dto.OAuth2ClientCreateFirstResponse": {
            "type": "object",
            "properties": {
                "allowed_scope": {
                    "type": "string",
                    "example": "read:user"
                },
                "client_id": {
                    "type": "string",
                    "example": "332974701238012989"
                },
                "client_secret": {
                    "type": "string",
                    "example": "ElBacv..."
                },
                "name": {
                    "type": "string",
                    "example": "Example Client"
                },
                "owner_id": {
                    "type": "string",
                    "example": "330559330522759168"
                }
            }
        },
        "dto.OAuth2ClientCreateRequest": {
            "type": "object",
            "properties": {
                "is_confidential": {
                    "type": "boolean",
                    "example": true
                },
                "name": {
                    "type": "string",
                    "example": "Example Client"
                }
            }
        },
        "dto.OAuth2ClientCreateResponse": {
            "type": "object",
            "properties": {
                "allowed_scope": {
                    "type": "string",
                    "example": "read:user"
                },
                "client_id": {
                    "type": "string",
                    "example": "332974701238012989"
                },
                "client_secret": {
                    "type": "string",
                    "example": "ElBacv..."
                },
                "name": {
                    "type": "string",
                    "example": "Example Client"
                },
                "owner_id": {
                    "type": "string",
                    "example": "330559330522759168"
                }
            }
        },
        "dto.OAuth2ClientGetResponse": {
            "type": "object",
            "properties": {
                "allowed_scope": {
                    "type": "string",
                    "example": "read:user"
                },
                "client_id": {
                    "type": "string",
                    "example": "332974701238012989"
                },
                "name": {
                    "type": "string",
                    "example": "Example Client"
                },
                "owner_id": {
                    "type": "string",
                    "example": "330559330522759168"
                }
            }
        },
...
        "standard.SwaggerBadRequestErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "invalid_request"
                },
                "error_description": {
                    "type": "string",
                    "example": "invalid field: require string but got int"
                },
                "metadata": {
                    "$ref": "#/definitions/standard.SwaggerMetadata"
                },
                "status": {
                    "type": "string",
                    "example": "error"
                }
            }
        },
...
        "standard.SwaggerForbiddenErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "forbidden"
                },
                "error_description": {
                    "type": "string",
                    "example": "not enough permission to access"
                },
                "metadata": {
                    "$ref": "#/definitions/standard.SwaggerMetadata"
                },
                "status": {
                    "type": "string",
                    "example": "error"
                }
            }
        },
...
        "standard.SwaggerMetadata": {
            "type": "object",
            "properties": {
                "request_id": {
                    "type": "string",
                    "example": "IIJORWQpIvDMzzNf"
                },
                "timestamp": {
                    "type": "string",
                    "example": "2024-10-23T13:52:29.459752901+07:00"
                }
            }
        },
        "standard.SwaggerNotFoundErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "not_found"
                },
                "error_description": {
                    "type": "string",
                    "example": "not found user with id 323979471029873"
                },
                "metadata": {
                    "$ref": "#/definitions/standard.SwaggerMetadata"
                },
                "status": {
                    "type": "string",
                    "example": "error"
                }
            }
        },
        "standard.SwaggerSuccessResponse-dto_OAuth2ClientCreateFirstResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/dto.OAuth2ClientCreateFirstResponse"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "standard.SwaggerSuccessResponse-dto_OAuth2ClientCreateResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/dto.OAuth2ClientCreateResponse"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "standard.SwaggerSuccessResponse-dto_OAuth2ClientGetResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/dto.OAuth2ClientGetResponse"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "standard.SwaggerUnauthorizedErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "forbidden"
                },
                "error_description": {
                    "type": "string",
                    "example": "require authentication"
                },
                "metadata": {
                    "$ref": "#/definitions/standard.SwaggerMetadata"
                },
                "status": {
                    "type": "string",
                    "example": "error"
                }
            }
        }
    }
}